test: 최단 경로 조회 단위 테스트 추가

+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -33,7 +33,7 @@ public class PathFinder {
         List<Station> stations = path.getVertexList();
         double distance = shortestPath.getPathWeight(starting, destination);
 
-        return new ShortestPathResponse(StationsResponse.of(new Stations(stations)), distance);
+        return new ShortestPathResponse(StationsResponse.of(new Stations(stations)), (int) distance);
     }
 
     private static WeightedMultigraph<Station, DefaultWeightedEdge> createStationGraph(List<Line> lines) {
